X Camera Rotation Works!
The only other problem is this solution requires the old mouse position
to be set at startup, a better solution for when the screen is resized
will have to be made.
Another problem is the Y rotation, which I guess is really the rotation
around the X Axis. Either way, this will have to be fixed as well
eventually.