Rotation is now working!
The main problem is that when I resize the screen the rotation gets inverted and all messed up. But, I believe this is a relatively small fix.
We are also doing rotation by rotating the individual target vector instead of trying to use vector coordinates, which I think are just not necessary.